fix map component structure
